fix(args): Eliminate the binary from arguments
For an executable the first argument is always the binary itself.
Because of this we have to skip it.

If we don't skip it, then it will be treated as another URL, therefore
`cosmic-player` will end up opening two URLs, one for the binary and one
for the actual file we want to play.

When opening multiple files `cosmic-player` will show the side panel /
playlist with the list of files.

The regression this commit fixed was introduced by `01be9152`.

Closes #96.